+++ tests/test_tie_point_grid.py
@@ -53,6 +53,7 @@ class Test_Tie_Point_Grid(unittest.TestCase):
                                  outFillVal=CRL.outFillVal,
                                  resamp_alg_calc=CRL.rspAlg_calc,
                                  tieP_filter_level=CRL.tieP_filter_level,
+                                 tieP_random_state=CRL.tieP_random_state,
                                  outlDetect_settings=dict(
                                      min_reliability=CRL.min_reliability,
                                      rs_max_outlier=CRL.rs_max_outlier,
@@ -175,6 +176,17 @@ class Test_Tie_Point_Grid(unittest.TestCase):
             arr_interp = self.TPG.to_interpolated_raster('ABS_SHIFT', 'Kriging', plot_result=True)
 
             assert isinstance(arr_interp, np.ndarray)
+
+    def test_random_state(self):
+        self.TPG.tieP_random_state = None
+        point_ids = [self.TPG.get_CoRegPoints_table()['POINT_ID'] for _ in range(2)]
+        assert not np.array_equal(point_ids[0], point_ids[1]), \
+            "Samples should not be identical when random state is None"
+
+        self.TPG.tieP_random_state = 0
+        point_ids = [self.TPG.get_CoRegPoints_table()['POINT_ID'] for _ in range(2)]
+        assert np.array_equal(point_ids[0], point_ids[1]), \
+            "Samples should be identical when random state is fixed"
 
 
 if __name__ == '__main__':